+++ backend/canisters/community/api/src/lib.rs
@@ -16,13 +16,6 @@ pub use updates::*;
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum EventsResponse {
     Success(types::EventsResponse),
-    UserNotInCommunity,
-    UserNotInChannel,
-    ChannelNotFound,
-    ThreadNotFound,
-    UserSuspended,
-    UserLapsed,
-    ReplicaNotUpToDateV2(types::TimestampMillis),
     Error(OCError),
 }