+++ backend/canisters/community/api/src/updates/cancel_invites.rs
@@ -15,9 +15,5 @@ pub struct Args {
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
-    NotAuthorized,
-    ChannelNotFound,
-    UserSuspended,
-    UserLapsed,
     Error(OCError),
 }