+++ backend/canisters/community/impl/src/updates/delete_messages.rs
@@ -27,8 +27,8 @@ async fn delete_messages(args: Args) -> Response {
     if args.as_platform_moderator.unwrap_or_default() && caller != user_index_canister_id {
         match lookup_user(caller, user_index_canister_id).await {
             Ok(Some(u)) if u.is_platform_moderator => {}
-            Ok(_) => return NotPlatformModerator,
-            Err(error) => return InternalError(format!("{error:?}")),
+            Ok(_) => return Error(OCErrorCode::InitiatorNotAuthorized.into()),
+            Err(error) => return Error(error.into()),
         }
     }