+++ backend/canisters/community/impl/src/updates/send_message.rs
@@ -65,7 +65,7 @@ fn c2c_bot_send_message(args: c2c_bot_send_message::Args) -> c2c_bot_send_messag
             BotPermissions::from_message_permission((&args.content).into()),
         )
     }) {
-        return c2c_bot_send_message::Response::NotAuthorized;
+        return c2c_bot_send_message::Response::Error(OCErrorCode::InitiatorNotAuthorized.into());
     }
 
     match mutate_state(|state| send_message_impl(args, Some(bot_caller), finalised, state)) {