+++ backend/canisters/group/api/src/updates/cancel_invites.rs
@@ -14,8 +14,5 @@ pub struct Args {
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
-    NotAuthorized,
-    UserSuspended,
-    UserLapsed,
     Error(OCError),
 }