+++ backend/canisters/group/api/src/updates/start_video_call_v2.rs
@@ -16,6 +16,5 @@ pub struct Args {
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
-    NotAuthorized,
     Error(OCError),
 }