+++ backend/canisters/local_user_index/impl/src/updates/bot_delete_channel.rs
@@ -39,10 +39,7 @@ async fn bot_delete_channel(args: Args) -> Response {
         Ok(response) => match response {
             community_canister::c2c_bot_delete_channel::Response::Success => Success,
             community_canister::c2c_bot_delete_channel::Response::Error(error) => Error(error),
-            community_canister::c2c_bot_delete_channel::Response::NotAuthorized => NotAuthorized,
-            community_canister::c2c_bot_delete_channel::Response::CommunityFrozen => Frozen,
-            community_canister::c2c_bot_delete_channel::Response::ChannelNotFound => ChannelNotFound,
         },
-        Err(error) => C2CError(error.reject_code() as i32, error.message().to_string()),
+        Err(error) => Error(error.into()),
     }
 }